+++ src/components/table/__spec__.js
@@ -48,14 +48,15 @@ describe('Table', () => {
     );
 
     instanceCustomSort = TestUtils.renderIntoDocument(
-      <Table className='baz'
-             onChange={ spy }
-             sortOrder='desc'
-             sortedColumn='name'
-             currentPage='10'
-             pageSize='25'
-             totalRecords='2500'
-        >
+      <Table
+        className='baz'
+        onChange={ spy }
+        sortOrder='desc'
+        sortedColumn='name'
+        currentPage='10'
+        pageSize='25'
+        totalRecords='2500'
+       >
         <TableRow>
           <TableHeader sortable={ true } name='name'/>
         </TableRow>
@@ -616,6 +617,28 @@ describe('Table', () => {
       instancePager.onPagination('2', '25');
 
       expect(spy).toHaveBeenCalledWith('pager', options);
+    });
+
+    describe('if an emitPageSizeChange callback was passed', () => {
+      it('runs the callback with the new page size', () => {
+        let callbackSpy = jasmine.createSpy();
+        let instanceCallBack = TestUtils.renderIntoDocument(
+          <Table
+            className="foo"
+            paginate={ true }
+            currentPage='1'
+            pageSize='10'
+            emitPageSizeChange={ callbackSpy }
+            totalRecords='100'
+            onChange={ spy }
+          >
+            foo
+          </Table>
+        );
+
+        instanceCallBack.onPagination('2', '50');
+        expect(callbackSpy).toHaveBeenCalledWith('50');
+      });
     });
   });