+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * `Date` field uses I18n for formats and sanitizes inputs for passing
 * `Decimal` component can prevent decimal value from exceeding the precision setting
+* `Table` components now accept an `emitPageSizeChange` callback function as a prop.
 
 # 0.26.0