+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 
 * `Date` field uses I18n for formats and sanitizes inputs for passing
 * `Decimal` component can prevent decimal value from exceeding the precision setting
+* `Table` components now accept an `onPageSizeChange` callback function as a prop.
+* `Pager` component now emits which element has changed.
 
 # 0.26.0