+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# 0.27.0
+
+## :warning: Breaking Change - Default colour for Pill component has changed. :warning:
+
+* The default behaviour for the Pill component was to previously set as `info`. This is now set as `default` which is a grey colour.
+* Ensure you check for any implementations of the Pill component where the `as` prop is not defined and set this to `as='info'`.
+
+## New Components
+
+* Subheader component created to be used with the Table and TableAjax components
+
+## Component Enhancements
+
+* `Date` field uses I18n for formats and sanitizes inputs for passing
+* `Decimal` component can prevent decimal value from exceeding the precision setting
+* `Table` components now accept an `onPageSizeChange` callback function as a prop.
+* `Pager` component now emits which element has changed.
+
 # 0.26.0
 
 ## New Components
@@ -9,9 +27,12 @@
 
 * `Content` now has additional display options to customise the alignment, to render inline with it's title and to customise the title's width.
 * `Link` component now has a prop of `iconAlign` to align icons to the right of the link's text.
+* `Pod` enter triggers edit function and edit element is keyboard accessible
 * `Row` component can now be given a size to control the size of the gutter using the prop `gutter` (eg. `extra-small`, `small`, `medium`, `large` or `extra-large`).
 * `Row` can enable `columnDivide` to add dividing lines between columns.
 * `ShowEditPod` requires a tab press to focus on the first field of the contained form rather than automatically focusing on the first field
+* `Tabs` enter triggers tab load and navigation tabs are keyboard accessible
+* `Tabs` focus state is given the same styles as hover state
 
 ## Minor Improvements